requirements.txt
django
mysqlclient
djangorestframework
Gunicorn
whitenoise==3.3.1
django-cors-headers
